refactor: Delete user and associated YouTube completions
This commit deletes a user and all associated YouTube completions from the database. The code changes in the user.actions.ts file remove the user entry and also delete all YouTube completions related to that user. This is necessary to ensure data consistency and clean up the database when a user is deleted.