refactor: Set button loading state in TaskList component
This commit sets the button loading state in the TaskList component. The code changes in TaskList.tsx add the 'setButtonLoading' function to set the loading state to true when the 'onMutate' function is called. It also sets the loading state to false when the 'onSettled' function is called. This improves the user experience by providing visual feedback to the user during the card purchase process.